+++ test/searchfield_test.dart
@@ -15,6 +15,7 @@ import 'dart:math';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:searchfield/searchfield.dart';
+
 import 'meta_data.dart';
 
 void main() {
@@ -432,7 +433,27 @@ void main() {
           ));
     });
   });
+  group('Scrollbar should be visible on suggestions', () {
+    testWidgets('Scrollbar should be visible by default',
+        (WidgetTester tester) async {
+      await tester.pumpWidget(_boilerplate(
+          child: SearchField(
+        key: const Key('searchfield'),
+        suggestions: ['ABC', 'DEF', 'GHI']
+            .map((e) => SearchFieldListItem<String>(e))
+            .toList(),
+        suggestionState: Suggestion.expand,
+      )));
 
+      final textField = find.byType(TextFormField);
+      expect(textField, findsOneWidget);
+      await tester.tap(textField);
+      await tester.enterText(textField, '');
+      await tester.pumpAndSettle();
+      expect(find.byType(ListView), findsOneWidget);
+      expect(find.byType(RawScrollbar), findsOneWidget);
+    });
+  });
   group('Suggestions should respect suggestionDirection', () {
     testWidgets(
         'suggestions should respect suggestionDirection: SuggestionDirection.up',
@@ -502,5 +523,104 @@ void main() {
     expect(listFinder.evaluate().length, 1);
     final suggestionOffsetNew = suggestionsRenderBox.localToGlobal(Offset.zero);
     expect(suggestionOffsetNew, equals(Offset(0.0, textFieldSize.height)));
+  });
+
+  testWidgets('enabled parameter should allow user to use the widget',
+      (WidgetTester tester) async {
+    await tester.pumpWidget(_boilerplate(
+        child: SearchField(
+      key: const Key('searchfield'),
+      enabled: true,
+      suggestions: ['ABC', 'DEF', 'GHI']
+          .map((e) => SearchFieldListItem<String>(e))
+          .toList(),
+    )));
+    final listFinder = find.byType(ListView);
+    expect(find.byType(TextFormField), findsOneWidget);
+    expect(listFinder, findsNothing);
+    await tester.tap(find.byType(TextFormField));
+    await tester.enterText(find.byType(TextFormField), '');
+    await tester.pumpAndSettle();
+    expect(listFinder, findsOneWidget);
+  });
+
+  testWidgets(
+      'enabled parameter set to false should not allow user to use the widget',
+      (WidgetTester tester) async {
+    await tester.pumpWidget(_boilerplate(
+        child: SearchField(
+      key: const Key('searchfield'),
+      enabled: false,
+      suggestions: ['ABC', 'DEF', 'GHI']
+          .map((e) => SearchFieldListItem<String>(e))
+          .toList(),
+    )));
+    final listFinder = find.byType(ListView);
+    expect(find.byType(TextFormField), findsOneWidget);
+    expect(listFinder, findsNothing);
+    await tester.tap(find.byType(TextFormField));
+    await tester.enterText(find.byType(TextFormField), '');
+    await tester.pumpAndSettle();
+    expect(listFinder, findsNothing);
+  });
+
+  testWidgets(
+      'Searchfield should not find suggestion when typed reversed for default',
+      (WidgetTester tester) async {
+    final controller = TextEditingController();
+    await tester.pumpWidget(_boilerplate(
+        child: SearchField(
+      key: const Key('searchfield'),
+      suggestions: ['ABC', 'DEF', 'GHI', 'JKL']
+          .map((e) => SearchFieldListItem<String>(e))
+          .toList(),
+      controller: controller,
+      suggestionState: Suggestion.expand,
+    )));
+    final listFinder = find.byType(ListView);
+    final textField = find.byType(TextFormField);
+    expect(textField, findsOneWidget);
+    expect(listFinder, findsNothing);
+    await tester.tap(textField);
+    await tester.enterText(textField, 'AB');
+    await tester.pumpAndSettle();
+    expect(listFinder, findsOneWidget);
+    expect(find.text('ABC'), findsOneWidget);
+    expect(listFinder.evaluate().length, 1);
+    await tester.enterText(textField, 'BA');
+    await tester.pumpAndSettle();
+    expect(listFinder, findsNothing);
+  });
+
+  testWidgets(
+      'Searchfield should find suggestion when typed reversed if we add custom comparator for it',
+      (WidgetTester tester) async {
+    final controller = TextEditingController();
+    await tester.pumpWidget(_boilerplate(
+        child: SearchField(
+      key: const Key('searchfield'),
+      comparator: (inputText, suggestionKey) =>
+          suggestionKey.contains(inputText.split('').reversed.join()),
+      suggestions: ['ABC', 'DEF', 'GHI', 'JKL']
+          .map((e) => SearchFieldListItem<String>(e))
+          .toList(),
+      controller: controller,
+      suggestionState: Suggestion.expand,
+    )));
+    final listFinder = find.byType(ListView);
+    final textField = find.byType(TextFormField);
+    expect(textField, findsOneWidget);
+    expect(listFinder, findsNothing);
+    await tester.tap(textField);
+    await tester.enterText(textField, 'AB');
+    await tester.pumpAndSettle();
+    expect(listFinder, findsOneWidget);
+    expect(find.text('ABC'), findsOneWidget);
+    expect(listFinder.evaluate().length, 1);
+    await tester.enterText(textField, 'BA');
+    await tester.pumpAndSettle();
+    expect(listFinder, findsOneWidget);
+    expect(find.text('ABC'), findsOneWidget);
+    expect(listFinder.evaluate().length, 1);
   });
 }